Create VideoProjectFile component
